requirements.txt
python-docx
Pillow
Flask
